requirements.txt
numpy
pandas
pygame
pytz
Requests
rich
TA-Lib
Flask
Flask-SocketIO
eventlet
passlib
gunicorn
bcrypt
python-dotenv
python-socketio
websocket-client